requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
langchain-core
langchain-community
langchain-chroma
langchain-groq
langchain-text-splitters
chromadb
pandas
pypdf
gradio
requests
